Add PoC and WiP solution for redirects
