+++ docs/.vuepress/enhanceApp.js
@@ -33,36 +33,8 @@ const buildActiveHeaderLinkHandler = () => {
   };
 };
 
-/**
- * The function iterates over the routes that point to the latest documentation
- * (that routes that do not have a declared version in the path URL) and adds
- * redirects to that pages but with an implicitly declared version in the URL
- * path (redirect from "/9.0/license-key/ to "/license-key").
- *
- * @param {*} router The Vue Router instance.
- * @param {string} latestVersion The latest version of the documentation.
- */
-const addLatestVersionRedirection = (router, latestVersion) => {
-  const versionedPathRegExp = /^\/(\d+\.\d+|next)\//;
-  const toNonVersionedRedirects = router.getRoutes()
-    .reduce((routes, { path }) => {
-      if (!versionedPathRegExp.test(path)) {
-        routes.push({
-          path: `/${latestVersion}${path}`,
-          redirect: path
-        });
-      }
-
-      return routes;
-    }, []);
-
-  toNonVersionedRedirects.forEach(versionedRedirect => router.addRoute(versionedRedirect));
-};
-
-export default ({ router, isServer, siteData }) => {
+export default ({ router, isServer }) => {
   if (!isServer) {
-    addLatestVersionRedirection(router, siteData.themeConfig.latestVersion);
-
     router.afterEach(buildRegisterCleaner(instanceRegister));
     router.afterEach(buildActiveHeaderLinkHandler());
   }
